+++ internal/backend/local/hook_state.go
@@ -34,30 +34,7 @@ type StateHook struct {
 	// and PersistInterval is ignored if this is nil.
 	Schemas *schemarepo.Schemas
 
-	intermediatePersist IntermediateStatePersistInfo
-}
-
-type IntermediateStatePersistInfo struct {
-	// RequestedPersistInterval is the persist interval requested by whatever
-	// instantiated the StateHook.
-	//
-	// Implementations of [IntermediateStateConditionalPersister] should ideally
-	// respect this, but may ignore it if they use something other than the
-	// passage of time to make their decision.
-	RequestedPersistInterval time.Duration
-
-	// LastPersist is the time when the last intermediate state snapshot was
-	// persisted, or the time of the first report for Terraform Core if there
-	// hasn't yet been a persisted snapshot.
-	LastPersist time.Time
-
-	// ForcePersist is true when Terraform CLI has receieved an interrupt
-	// signal and is therefore trying to create snapshots more aggressively
-	// in anticipation of possibly being terminated ungracefully.
-	// [IntermediateStateConditionalPersister] implementations should ideally
-	// persist every snapshot they get when this flag is set, unless they have
-	// some external information that implies this shouldn't be necessary.
-	ForcePersist bool
+	intermediatePersist statemgr.IntermediateStatePersistInfo
 }
 
 var _ terraform.Hook = (*StateHook)(nil)
@@ -129,48 +106,8 @@ func (h *StateHook) Stopping() {
 }
 
 func (h *StateHook) shouldPersist() bool {
-	if m, ok := h.StateMgr.(IntermediateStateConditionalPersister); ok {
+	if m, ok := h.StateMgr.(statemgr.IntermediateStateConditionalPersister); ok {
 		return m.ShouldPersistIntermediateState(&h.intermediatePersist)
 	}
-	return DefaultIntermediateStatePersistRule(&h.intermediatePersist)
-}
-
-// DefaultIntermediateStatePersistRule is the default implementation of
-// [IntermediateStateConditionalPersister.ShouldPersistIntermediateState] used
-// when the selected state manager doesn't implement that interface.
-//
-// Implementers of that interface can optionally wrap a call to this function
-// if they want to combine the default behavior with some logic of their own.
-func DefaultIntermediateStatePersistRule(info *IntermediateStatePersistInfo) bool {
-	return info.ForcePersist || time.Since(info.LastPersist) >= info.RequestedPersistInterval
-}
-
-// IntermediateStateConditionalPersister is an optional extension of
-// [statemgr.Persister] that allows an implementation to tailor the rules for
-// whether to create intermediate state snapshots when Terraform Core emits
-// events reporting that the state might have changed.
-//
-// For state managers that don't implement this interface, [StateHook] uses
-// a default set of rules that aim to be a good compromise between how long
-// a state change can be active before it gets committed as a snapshot vs.
-// how many intermediate snapshots will get created. That compromise is subject
-// to change over time, but a state manager can implement this interface to
-// exert full control over those rules.
-type IntermediateStateConditionalPersister interface {
-	// ShouldPersistIntermediateState will be called each time Terraform Core
-	// emits an intermediate state event that is potentially eligible to be
-	// persisted.
-	//
-	// The implemention should return true to signal that the state snapshot
-	// most recently provided to the object's WriteState should be persisted,
-	// or false if it should not be persisted. If this function returns true
-	// then the receiver will see a subsequent call to
-	// [statemgr.Persister.PersistState] to request persistence.
-	//
-	// The implementation must not modify anything reachable through the
-	// arguments, and must not retain pointers to anything reachable through
-	// them after the function returns. However, implementers can assume that
-	// nothing will write to anything reachable through the arguments while
-	// this function is active.
-	ShouldPersistIntermediateState(info *IntermediateStatePersistInfo) bool
+	return statemgr.DefaultIntermediateStatePersistRule(&h.intermediatePersist)
 }
