+++ internal/states/remote/state.go
@@ -11,7 +11,6 @@ import (
 
 	uuid "github.com/hashicorp/go-uuid"
 
-	"github.com/hashicorp/terraform/internal/backend/local"
 	"github.com/hashicorp/terraform/internal/schemarepo"
 	"github.com/hashicorp/terraform/internal/states"
 	"github.com/hashicorp/terraform/internal/states/statefile"
@@ -50,7 +49,7 @@ type State struct {
 
 var _ statemgr.Full = (*State)(nil)
 var _ statemgr.Migrator = (*State)(nil)
-var _ local.IntermediateStateConditionalPersister = (*State)(nil)
+var _ statemgr.IntermediateStateConditionalPersister = (*State)(nil)
 
 // statemgr.Reader impl.
 func (s *State) State() *states.State {
@@ -227,12 +226,12 @@ func (s *State) PersistState(schemas *schemarepo.Schemas) error {
 	return nil
 }
 
-// ShouldPersistIntermediateState implements local.IntermediateStateConditionalPersister
-func (s *State) ShouldPersistIntermediateState(info *local.IntermediateStatePersistInfo) bool {
+// ShouldPersistIntermediateState implements statemgr.IntermediateStateConditionalPersister
+func (s *State) ShouldPersistIntermediateState(info *statemgr.IntermediateStatePersistInfo) bool {
 	if s.DisableIntermediateSnapshots {
 		return false
 	}
-	return local.DefaultIntermediateStatePersistRule(info)
+	return statemgr.DefaultIntermediateStatePersistRule(info)
 }
 
 // Lock calls the Client's Lock method if it's implemented.
